[rollout] refactor: single home for diagnostics and error categorization
Backend diagnostics now persist only via the diagnostics.json sidecar;
the parallel result_extra_fields channel into the trajectory document's
extra block duplicated the same five-key payload and is dropped.
categorize_exception moves to osmosis_ai.rollout.utils.errors, replacing
the byte-identical copies in the local backend and harbor diagnostics.
