+++ osmosis_ai/rollout/trajectory/converter.py
@@ -45,7 +45,6 @@ def convert_sample_to_trajectory(
     request_label: str | None = None,
     request_metadata: dict[str, Any] | None = None,
     request_extra_fields: dict[str, Any] | None = None,
-    result_extra_fields: dict[str, Any] | None = None,
     report: SampleReport | None = None,
     default_model_name: str | None = None,
     unmatched_sample_reports: Mapping[str, SampleReport] | None = None,
@@ -55,7 +54,6 @@ def convert_sample_to_trajectory(
     ``report`` carries controller-reported per-call metrics (see
     ``report.py``); ``unmatched_sample_reports`` are entries that could
     not be attributed to the sample, preserved under ``extra``.
-    ``result_extra_fields`` carries the backend's execution diagnostics.
     """
     if sample.trajectory_messages is None:
         raise ValueError("Sample has no trajectory-compatible messages")
@@ -84,7 +82,6 @@ def convert_sample_to_trajectory(
             request_label=request_label,
             request_metadata=request_metadata,
             request_extra_fields=request_extra_fields,
-            result_extra_fields=result_extra_fields,
             unmatched_llm_call_metrics=unmatched_llm_call_metrics,
             unmatched_sample_reports=unmatched_reports,
         ),
@@ -389,7 +386,6 @@ def _compose_extra(
     request_label: str | None,
     request_metadata: dict[str, Any] | None,
     request_extra_fields: dict[str, Any] | None,
-    result_extra_fields: dict[str, Any] | None,
     unmatched_llm_call_metrics: list[dict[str, Any]] | None,
     unmatched_sample_reports: dict[str, Any] | None,
 ) -> dict[str, Any]:
@@ -402,7 +398,6 @@ def _compose_extra(
         "sample_extra_fields": sample.extra_fields or None,
         "request_metadata": request_metadata,
         "request_extra_fields": request_extra_fields,
-        "result_extra_fields": result_extra_fields,
         "unmatched_llm_call_metrics": unmatched_llm_call_metrics,
         "unmatched_sample_reports": unmatched_sample_reports,
     }
